Update lit tests
